fix(clippy): clean both-chip clippy for the 0.5.0 release (Area H)
- safety.rs: `TIMER_CLOCK_HZ.is_multiple_of(1_000_000)` → `% 1_000_000 == 0`
  (`is_multiple_of` is const-stable only since 1.87; this keeps the declared
  MSRV 1.85 honest — clippy::incompatible_msrv).
- pwm.rs: drop the redundant `u32 -> u32` casts in the period read-back.
- spi.rs: `SpiError::kind` uses a total match instead of match+return+unreachable.
- dma.rs: gate the `Transfer` cache-only fields (`dst_addr`/`bytes`) `allow(dead_code)`
  off chip-ws63 (they feed the ws63-gated cache maintenance).

Both chips are now clippy-clean (riscv target); 309 host tests pass; fmt clean.

+++ src/pwm.rs
@@ -182,7 +182,7 @@ impl<'d> PwmChannel<'d> {
             7 => (r.pwm_freq_l7().read().bits(), r.pwm_freq_h7().read().bits()),
             _ => unreachable!(),
         };
-        (lo as u32 & 0xFFFF) | ((hi as u32 & 0xFFFF) << 16)
+        (lo & 0xFFFF) | ((hi & 0xFFFF) << 16)
     }
 
     /// Configure this channel's period and duty cycle, bringing up the PWM clock
